refactor and remove node-list-tree-search
