+++ package.json
@@ -42,27 +42,27 @@
     "@material-ui/lab": "^4.0.0-alpha.58",
     "@quantumblack/kedro-ui": "1.1.4",
     "batching-toposort": "^1.2.0",
-    "classnames": "^2.2.6",
+    "classnames": "^2.3.1",
     "d3-fetch": "^2.0.0",
     "d3-interpolate": "^2.0.1",
-    "d3-interpolate-path": "^2.2.1",
-    "d3-scale": "^3.2.3",
+    "d3-interpolate-path": "^2.2.3",
+    "d3-scale": "^3.3.0",
     "d3-selection": "^2.0.0",
-    "d3-shape": "^2.0.0",
+    "d3-shape": "^2.1.0",
     "d3-transition": "^2.0.0",
     "d3-zoom": "^2.0.0",
     "dagre": "git+https://github.com/richardwestenra/dagre.git#manual-ranking",
     "deepmerge": "^4.2.2",
     "fetch-mock": "^9.11.0",
-    "highlight.js": "^10.5.0",
+    "highlight.js": "^10.7.3",
     "kiwi.js": "^1.1.2",
     "lodash.clonedeep": "^4.5.0",
     "plotly.js-cartesian-dist": "^1.58.4",
     "react-custom-scrollbars": "^4.2.1",
     "react-plotly.js": "^2.5.1",
-    "react-redux": "^7.2.2",
-    "react-test-renderer": "^17.0.1",
-    "redux": "^4.0.5",
+    "react-redux": "^7.2.4",
+    "react-test-renderer": "^17.0.2",
+    "redux": "^4.1.0",
     "redux-mock-store": "^1.5.4",
     "redux-thunk": "^2.3.0",
     "redux-watch": "^1.2.0",
@@ -76,29 +76,29 @@
     "react-dom": "^16.8.6"
   },
   "devDependencies": {
-    "@babel/cli": "^7.12.10",
-    "@babel/core": "^7.12.10",
-    "@babel/plugin-proposal-class-properties": "^7.12.1",
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.4",
-    "@testing-library/user-event": "^12.6.3",
+    "@babel/cli": "^7.14.5",
+    "@babel/core": "^7.14.6",
+    "@babel/plugin-proposal-class-properties": "^7.14.5",
+    "@testing-library/jest-dom": "^5.14.1",
+    "@testing-library/react": "^11.2.7",
+    "@testing-library/user-event": "^12.8.3",
     "canvas": "^2.7.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.6",
     "husky": "^4.3.8",
-    "lint-staged": "^10.5.3",
+    "lint-staged": "^10.5.4",
     "node-sass": "^5.0.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.3.0",
+    "prettier": "^2.3.1",
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
-    "react-scripts": "^4.0.1",
-    "snyk": "^1.623.0",
-    "stylelint": "^13.9.0",
+    "react-scripts": "^4.0.3",
+    "snyk": "^1.639.0",
+    "stylelint": "^13.13.1",
     "stylelint-config-prettier": "^8.0.2",
     "stylelint-order": "^4.1.0",
-    "stylelint-prettier": "^1.1.2",
-    "webpack-cli": "^4.4.0",
+    "stylelint-prettier": "^1.2.0",
+    "webpack-cli": "^4.7.2",
     "workerize-loader": "^1.3.0"
   },
   "husky": {
