+++ src/components/node-list/node-list-tree.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
 
 import { makeStyles, withStyles } from '@material-ui/core/styles';
@@ -18,7 +18,7 @@ import {
 } from '../../selectors/modular-pipelines';
 import {
   getNestedModularPipelines,
-  getTreeSearchValueCount,
+  getFilteredTreeItems,
 } from './node-list-items';
 import {
   getGroupedNodes,
@@ -56,8 +56,42 @@ const TreeListProvider = ({
   onItemMouseEnter,
   onItemMouseLeave,
   nodeTypeIDs,
+  searching,
 }) => {
   const classes = useStyles();
+
+  const [expandedPipelines, setExpandedPipelines] = useState([]);
+
+  useEffect(() => {
+    const filteredTreeItems =
+      searchValue !== ''
+        ? getFilteredTreeItems({
+            nodes,
+            modularPipelines,
+            nodeSelected,
+            searchValue,
+            modularPipelineIds,
+            nodeModularPipelines,
+            nodeTypeIDs,
+          })
+        : [];
+
+    let expandedModularPipelines = [];
+
+    searchValue !== '' &&
+      filteredTreeItems.forEach((modularPipeline) =>
+        expandedModularPipelines.push(modularPipeline.id)
+      );
+    setExpandedPipelines(expandedModularPipelines);
+  }, [
+    searchValue,
+    nodes,
+    modularPipelines,
+    nodeSelected,
+    modularPipelineIds,
+    nodeModularPipelines,
+    nodeTypeIDs,
+  ]);
 
   const treeData = getNestedModularPipelines({
     nodes,
@@ -69,21 +103,42 @@ const TreeListProvider = ({
     nodeTypeIDs,
   });
 
-  const searchTreeCount = getTreeSearchValueCount({ treeData, searchValue });
-
   const onItemClick = (item) => {
     if (!isModularPipelineType(item.type)) {
       onToggleNodeSelected(item.id);
     }
   };
 
+  const renderModularPipelines = (treeData) =>
+    treeData.children.length > 0 &&
+    treeData.children.map((node) =>
+      renderTree(
+        node,
+        onItemMouseEnter,
+        onItemMouseLeave,
+        onItemChange,
+        onItemClick
+      )
+    );
+
+  const renderChildNodes = (treeData) =>
+    treeData.nodes.map((node) => (
+      <NodeListTreeItem
+        data={node}
+        onItemMouseEnter={onItemMouseEnter}
+        onItemMouseLeave={onItemMouseLeave}
+        onItemChange={onItemChange}
+        onItemClick={onItemClick}
+        key={node.id}
+      />
+    ));
+
   const renderTree = (
     rowData,
     onItemMouseEnter,
     onItemMouseLeave,
     onItemChange,
-    onItemClick,
-    searchTreeCount
+    onItemClick
   ) => (
     <NodeListTreeItem
       data={rowData}
@@ -91,62 +146,32 @@ const TreeListProvider = ({
       onItemMouseLeave={onItemMouseLeave}
       onItemChange={onItemChange}
       onItemClick={onItemClick}
-      key={rowData.id}
-      searchCount={searchTreeCount[rowData.id]}>
-      {rowData.children.length > 0 &&
-        rowData.children.map((node) =>
-          renderTree(
-            node,
-            onItemMouseEnter,
-            onItemMouseLeave,
-            onItemChange,
-            onItemClick,
-            searchTreeCount
-          )
-        )}
+      key={rowData.id}>
+      {renderModularPipelines(rowData)}
 
       {/* render set of node elements in that modular pipeline */}
-      {rowData.nodes.map((node) => (
-        <NodeListTreeItem
-          data={node}
-          onItemMouseEnter={onItemMouseEnter}
-          onItemMouseLeave={onItemMouseLeave}
-          onItemChange={onItemChange}
-          onItemClick={onItemClick}
-          key={node.id}
-        />
-      ))}
+      {renderChildNodes(rowData)}
     </NodeListTreeItem>
   );
 
-  return (
+  return searching ? (
+    <StyledTreeView
+      className={classes.root}
+      defaultCollapseIcon={<ExpandMoreIcon />}
+      defaultExpandIcon={<ChevronRightIcon />}
+      expanded={expandedPipelines}>
+      {/* render set of modular pipelines in the main pipeline */}
+      {renderModularPipelines(treeData)}
+      {/* render set of node elements in the main pipeline */}
+      {renderChildNodes(treeData)}
+    </StyledTreeView>
+  ) : (
     <StyledTreeView
       className={classes.root}
       defaultCollapseIcon={<ExpandMoreIcon />}
       defaultExpandIcon={<ChevronRightIcon />}>
-      {treeData.children.length > 0 &&
-        treeData.children.map((node) =>
-          renderTree(
-            node,
-            onItemMouseEnter,
-            onItemMouseLeave,
-            onItemChange,
-            onItemClick,
-            searchTreeCount
-          )
-        )}
-
-      {/* render set of node elements in the main pipeline */}
-      {treeData.nodes.map((node) => (
-        <NodeListTreeItem
-          data={node}
-          onItemMouseEnter={onItemMouseEnter}
-          onItemMouseLeave={onItemMouseLeave}
-          onItemChange={onItemChange}
-          onItemClick={onItemClick}
-          key={node.id}
-        />
-      ))}
+      {renderModularPipelines(treeData)}
+      {renderChildNodes(treeData)}
     </StyledTreeView>
   );
 };
